Continue split code into modules and make it more PEP8 compliant
